Rename Header.txt to index.tsx and change imports
